+++ compleet-client/src/bindings/api/window.rs
@@ -7,7 +7,7 @@ use mlua::{
 };
 
 /// Binding to `vim.api.nvim_win_close`.
-pub fn win_close(lua: &Lua, winid: u32, force: bool) -> LuaResult<()> {
+pub fn win_close(lua: &Lua, winid: u16, force: bool) -> LuaResult<()> {
     super::api(lua)?
         .get::<&str, LuaFunction>("nvim_win_close")?
         .call((winid, force))
@@ -25,7 +25,7 @@ pub fn win_get_cursor(lua: &Lua, winid: u16) -> LuaResult<(u16, u16)> {
 /// Binding to `vim.api.nvim_win_get_option`
 pub fn win_get_option<'lua, V: FromLua<'lua>>(
     lua: &'lua Lua,
-    winid: u32,
+    winid: u16,
     name: &str,
 ) -> LuaResult<V> {
     super::api(lua)?
@@ -34,7 +34,7 @@ pub fn win_get_option<'lua, V: FromLua<'lua>>(
 }
 
 /// Binding to `vim.api.nvim_win_get_position`
-pub fn win_get_position(lua: &Lua, winid: u32) -> LuaResult<(u16, u16)> {
+pub fn win_get_position(lua: &Lua, winid: u16) -> LuaResult<(u16, u16)> {
     let position = super::api(lua)?
         .get::<&str, LuaFunction>("nvim_win_get_position")?
         .call::<_, Table>(winid)?;
@@ -57,14 +57,14 @@ pub fn win_get_height(lua: &Lua, winid: u16) -> LuaResult<u16> {
 }
 
 /// Binding to `vim.api.nvim_win_hide`.
-pub fn win_hide(lua: &Lua, winid: u32) -> LuaResult<()> {
+pub fn win_hide(lua: &Lua, winid: u16) -> LuaResult<()> {
     super::api(lua)?.get::<&str, LuaFunction>("nvim_win_hide")?.call(winid)
 }
 
 /// Binding to `vim.api.nvim_win_set_cursor`
 pub fn win_set_cursor(
     lua: &Lua,
-    winid: u32,
+    winid: u16,
     row: u32,
     col: u32,
 ) -> LuaResult<()> {
@@ -76,7 +76,7 @@ pub fn win_set_cursor(
 /// Binding to `vim.api.nvim_win_set_option`
 pub fn win_set_option<'lua, V: ToLua<'lua>>(
     lua: &'lua Lua,
-    winid: u32,
+    winid: u16,
     name: &str,
     value: V,
 ) -> LuaResult<()> {
